+++ lib/predicate/nodes/lte.rb
@@ -10,5 +10,11 @@ class Predicate
       left.evaluate(tuple) <= right.evaluate(tuple)
     end
 
+    def assert!(tuple, asserter = Asserter.new)
+      l, r = left.evaluate(tuple), right.evaluate(tuple)
+      asserter.assert_operator(l, :<=, r)
+      l
+    end
+
   end
 end