+++ lib/predicate/nodes/neq.rb
@@ -16,5 +16,16 @@ class Predicate
       l
     end
 
+    def to_hashes
+      hash = if left.identifier? && right.literal? && !right.has_placeholder?
+        { left.name => right.value }
+      elsif right.identifier? && left.literal? && !left.has_placeholder?
+        { right.name => left.value }
+      else
+        super
+      end
+      [ {}, hash ]
+    end
+
   end
 end