+++ lib/predicate/nodes/tautology.rb
@@ -37,5 +37,9 @@ class Predicate
       {}
     end
 
+    def to_hashes
+      [{}, {}]
+    end
+
   end
 end